+++ workout-tracker/src/views/UserPage/UserPage.css
@@ -15,6 +15,11 @@
   display: flex;
   flex-direction: column;
   background-color: #f2f2f2;
+  flex:1;
+}
+
+.dashboard-view {
+  flex:4;
 }
 
 .nav-items {
